split out `Abstract` and `Interpret`, implement rotate + reflect
