+++ src/Swarm/Game/World/Eval.hs
@@ -18,9 +18,10 @@ import Swarm.Game.Failure (Asset (..), AssetData (..), LoadingFailure (..), Syst
 import Swarm.Game.ResourceLoading (getDataDirSafe)
 import Swarm.Game.Terrain (TerrainType (..))
 import Swarm.Game.World (WorldFun (..))
-import Swarm.Game.World.Compile
+import Swarm.Game.World.Abstract (bracket)
 import Swarm.Game.World.Coords (Coords)
 import Swarm.Game.World.Gen (Seed)
+import Swarm.Game.World.Interpret (interpBTerm)
 import Swarm.Game.World.Parse (parseWExp, runParser)
 import Swarm.Game.World.Syntax
 import Swarm.Game.World.Typecheck
